+++ src/components/Workflow/nodes/index.ts
@@ -3,48 +3,57 @@ import type { Node, NodeTypes, BuiltInNode } from "@xyflow/react";
 import { AnalysisInputNode } from "./AnalysisInputNode";
 import { AnalysisOutputNode } from "./AnalysisOutputNode";
 import { AnalysisNode } from "./AnalysisNode";
+import { NewAnalysisNode } from "./NewAnalysisNode";
 
 export type AnalysisInputNode = Node<
     {
-        label?: string;
-        inputData?: DatawolfIO | { id: string; title: string };
-        type?: "dataset" | "hazard" | "dfr3_mapping";
+        label: string;
+        inputData: DatawolfIO | { id: string; title: string; dataId: string };
+        stepID: string;
+        type: "dataset" | "hazard" | "dfr3_mapping";
+        analysisName: string;
     },
     "analysis-input"
 >;
 
 export type AnalysisOutputNode = Node<
     {
-        label?: string;
-        outputData?: DatawolfIO | { id: string; title: string };
+        label: string;
+        outputData: DatawolfIO;
+        stepID: string;
+        analysisName: string;
     },
     "analysis-output"
 >;
 
 export type AnalysisNode = Node<
     {
-        label?: string;
+        label: string;
+        name: string;
         stepData?: DatawolfWorkflowFileStep;
+        toolID?: string;
     },
     "analysis"
 >;
 
-export type AppNode = BuiltInNode | AnalysisInputNode | AnalysisOutputNode | AnalysisNode;
+export type NewAnalysisNode = Node<
+    {
+        label: string;
+        name: string;
+        inputHandles: { id: string; label: string; dataId: string; type: string }[];
+        outputHandles: { id: string; label: string; dataId: string; type: string }[];
+        stepData?: DatawolfWorkflowFileStep;
+        tool?: DatawolfWorkflowTool;
+    },
+    "new-analysis-node"
+>;
 
-const position = { x: 0, y: 0 };
-
-export const initialNodes: AppNode[] = [
-    { id: "inpt1", type: "analysis-input", position: position, data: { label: "Bridge Dataset" } },
-    { id: "inpt2", type: "analysis-input", position: position, data: { label: "Hazard Type" } },
-    { id: "inpt3", type: "analysis-input", position: position, data: { label: "DFR3 Mapping Set" } },
-    { id: "analysis1", type: "analysis", position: position, data: { label: "Bridge Damage" } },
-    { id: "output1", type: "analysis-output", position: position, data: { label: "Type 1" } },
-    { id: "output2", type: "analysis-output", position: position, data: { label: "Type 2" } }
-];
+export type AppNode = BuiltInNode | AnalysisInputNode | AnalysisOutputNode | AnalysisNode | NewAnalysisNode;
 
 export const nodeTypes = {
     "analysis-input": AnalysisInputNode,
     "analysis-output": AnalysisOutputNode,
-    "analysis": AnalysisNode
+    "analysis": AnalysisNode,
+    "new-analysis-node": NewAnalysisNode
     // Add any of your custom nodes here!
 } satisfies NodeTypes;