+++ CHANGELOG.md
@@ -13,4 +13,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Github action for building container [#2](https://github.com/IN-CORE/incore-studio/issues/2)
 - Keycloak authentication to every route [#3](https://github.com/IN-CORE/incore-studio/issues/3)
 - Create home page that lists all the projects [#4](https://github.com/IN-CORE/incore-studio/issues/4)
+- Add Workflow Editor component. [#15](https://github.com/IN-CORE/incore-studio/issues/15)
+- Add "Add Next" and "Add Previous" analysis in analysis node. [#6](https://github.com/IN-CORE/incore-studio/issues/6)
 - Create navbar components [#12](https://github.com/IN-CORE/incore-studio/issues/12)